Staging

+++ app/Http/Controllers/Dashboard/UserController.php
@@ -28,26 +28,20 @@ class UserController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
+        $validated = $request->validate([
             'nama' => 'required',
             'email' => 'required|email|unique:users,email',
-            'tanggal_lahir' => 'required|date',
+            'tanggal_lahir' => 'nullable|date',
             'role_id' => 'required|exists:roles,id',
         ]);
 
-        $data = [
-            'nama' => $request->input('nama'),
-            'email' => $request->input('email'),
-            'role_id' => $request->input('role_id'),
-        ];
-
         if($request->has('tanggal_lahir')) {
-            $data['tanggal_lahir'] = $request->input('tanggal_lahir');
-            $data['password'] = bcrypt($request->input('tanggal_lahir'));
+            $tanggal_lahir = $request->input('tanggal_lahir');
+            $validated['tanggal_lahir'] = $tanggal_lahir;
+            $validated['password'] = bcrypt($tanggal_lahir);
         }
 
-        // Tambahkan data user
-        $user = User::create($data);
+        $user = User::create($validated); // Tambahkan data user
 
         // Cek role_id
         if ($user->role_id == ($user->role->nama == 'dosen')) {
@@ -63,8 +57,6 @@ class UserController extends Controller
                 'user_id' => $user->id,
             ]);
         }
-
-        $user->save();
 
         session()->flash('success', 'Data User berhasil ditambah');
 
@@ -84,6 +76,7 @@ class UserController extends Controller
         $request->validate([
             'nama' => 'required',
             'email' => 'required|email|unique:users,email,' . $user->id,
+            'password' => 'nullable|min:8',
             'tanggal_lahir' => 'required|date',
             'role_id' => 'required',
         ]);
@@ -107,7 +100,7 @@ class UserController extends Controller
             $user->update([
                 'nama' => $request->nama,
                 'email' => $request->email,
-                'password' => $request->$password,
+                'password' => $password,
                 'role_id' => $request->role_id,
             ]);
         }
@@ -124,7 +117,10 @@ class UserController extends Controller
         }
 
         if ($request->role_id == Role::where('nama', 'mahasiswa')->first()->id){
-            if (!$user->mahasiswa) {
+            if (isset($user->mahasiswa->tugas_akhir)) {
+                return redirect()->back()->with('error', 'Data mahasiswa terkait dengan tugas akhir');
+            }
+            elseif (!$user->mahasiswa) {
                 // Buat data baru pada tabel mahasiswa
                 Mahasiswa::create([
                     'user_id' => $user->id,
@@ -142,15 +138,17 @@ class UserController extends Controller
     {
         // Menghapus user
         if ($user->dosen){
+            if($user->dosen->dosen_penguji || $user->dosen->dosen_pembimbing){
+                return redirect()->back()->with('error', 'Tidak bisa menghapus data dosen yang sudah terkait sebagai dosen pembimbing atau dosen penguji.');
+            }
             $user->dosen->delete();
-            // Menghapus data dosen_pengujis terkait
-            $user->dosen->dosen_pengujis->delete();
-            // Menghapus data dosen_pembimbings terkait
-            $user->dosen->dosen_pembimbings->delete();
-            // Menghapus user
             $user->delete();
         } else if ($user->mahasiswa){
+            if($user->mahasiswa->tugas_akhir){
+                return redirect()->back()->with('error', 'Tidak bisa menghapus data dosen yang sudah terkait sebagai dosen pembimbing atau dosen penguji.');
+            }
             $user->mahasiswa->delete();
+            $user->delete();
         } else {
             $user->delete();
         }
